% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aco.operator.R
\name{aco.operator}
\alias{aco.operator}
\title{Run ant colony optimisation for automated model selection}
\usage{
aco.operator(
  dat,
  no.ants,
  max.iter,
  rho,
  initial.phi,
  lower.limit.phi,
  upper.limit.phi,
  alpha.value,
  elitism.percentage,
  sig.diff,
  search.space,
  no.cores,
  thetalower,
  crse,
  cshrink,
  cadd,
  cprop,
  ccorr,
  penalty.type,
  penalty.value,
  foldername,
  filename,
  ...
)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data to be used for model fitting.}

\item{no.ants}{An integer specifying the number of ants to use. Default is 20.}

\item{max.iter}{An integer specifying the maximum number of iterations. Default is 15.}

\item{rho}{A numeric value for the pheromone evaporation rate. Default is 0.2.}

\item{initial.phi}{A numeric value for the initial pheromone level. Default is 1.}

\item{lower.limit.phi}{A numeric value for the lower limit of the pheromone level. Default is 1.}

\item{upper.limit.phi}{A numeric value for the upper limit of the pheromone level. Default is Inf.}

\item{alpha.value}{A numeric value for the pheromone influence parameter. Default is 1.}

\item{elitism.percentage}{A numeric value specifying the percentage of elitism. Default is 1/3}

\item{sig.diff}{The significance difference threshold for ranking.}

\item{search.space}{An integer specifying the search space type (1 for standard).}

\item{no.cores}{The number of CPU cores to use for parallel processing.}

\item{thetalower}{A named vector specifying lower bounds for model parameters.}

\item{crse}{Constraint value for the relative standard error (RSE).}

\item{cshrink}{Constraint value for the shrinkage.}

\item{cadd}{Constraint value for the additive error.}

\item{cprop}{Constraint value for the proportional error.}

\item{ccorr}{Constraint value for the correlation between parameters.}

\item{penalty.type}{The type of penalty to apply during the fitness evaluation:
\itemize{
\item \code{penalty.type = 1}: Considers constraints on RSE.
\item \code{penalty.type = 2}: Considers constraints on RSE and shrinkage.
\item \code{penalty.type = 3}: Considers constraints on RSE, shrinkage, and sigma values (variance of residual error).
}}

\item{penalty.value}{The value of the penalty to apply during the fitness evaluation.}

\item{foldername}{The name of the folder for output storage. Defaults to "test".}

\item{filename}{The name of the output file. Defaults to "test".}

\item{...}{Additional arguments passed to the model fitting function.}
}
\value{
A list include the best solution and ACO running history
}
\description{
Perform an ant colony optimisation (ACO) algorithm to select the best model for pharmacokinetic modelling
}
\examples{
\dontrun{
d1<-pheno_sd
inits.out<-getppkinit(dat = d1,runnpd = 0)
autosetinit(dat = d1,
           inits= inits.out$Recommended_initial_estimates)
result <- aco.operator(
    dat = d1,
    no.ants = 6,
    max.iter = 6,
    rho = 0.2,
    sig.diff = 1,
    search.space = 1,
    no.cores = 4,
    thetalower = c(vp = 1, vp2 = 1),
    control = saemControl(
    seed = 1234,
    print = 5,
    nBurn = 20,
    nEm = 30,
    logLik = TRUE,
    rxControl = rxControl(cores = 4, maxsteps = 70000)
   ),
   table = tableControl(cwres = TRUE),
   filename = "pheno_sd",
   foldername = "pheno_sd"
 )
result
}

}
