% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosetinit.R
\name{autosetinit}
\alias{autosetinit}
\title{Automatically set initial values for model parameters}
\usage{
autosetinit(
  dat,
  inits,
  input_vmax = NULL,
  input_km = NULL,
  input_cl = NULL,
  input_vc = NULL,
  input_vc2cpt = NULL,
  input_vp2cpt = NULL,
  input_q2cpt = NULL,
  input_vc3cpt = NULL,
  input_vp3cpt = NULL,
  input_vp23cpt = NULL,
  input_q3cpt = NULL,
  input_q23cpt = NULL,
  input_prop.err = NULL,
  input_add.err = NULL,
  input_logn.sd = NULL,
  input_eta.cl = NULL,
  input_eta.vc = NULL,
  input_eta.vp = NULL,
  input_eta.vp2 = NULL,
  input_eta.q = NULL,
  input_eta.q2 = NULL,
  input_eta.vmax = NULL,
  input_eta.km = NULL,
  input_etarsquare = NULL
)
}
\arguments{
\item{dat}{A data frame containing the data to be used for setting initial values.}

\item{inits}{A data frame containing the initial estimates of the parameters, which is obtained from \code{nlmixr2autoinit}}

\item{input_vmax}{Optional. Initial value for Vmax. If not provided, it will be set based on \code{inits}.}

\item{input_km}{Optional. Initial value for Km. If not provided, it will be set based on \code{inits}.}

\item{input_cl}{Optional. Initial value for clearance (CL). If not provided, it will be set based on \code{inits}.}

\item{input_vc}{Optional. Initial value for volume of distribution (Vc). If not provided, it will be set based on \code{inits}.}

\item{input_vc2cpt}{Optional. Initial value for Vc in a two-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_vp2cpt}{Optional. Initial value for Vp in a two-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_q2cpt}{Optional. Initial value for Q in a two-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_vc3cpt}{Optional. Initial value for Vc in a three-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_vp3cpt}{Optional. Initial value for Vp in a three-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_vp23cpt}{Optional. Initial value for Vp2 in a three-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_q3cpt}{Optional. Initial value for Q in a three-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_q23cpt}{Optional. Initial value for Q2 in a three-compartment model. If not provided, it will be set based on \code{inits}.}

\item{input_prop.err}{Optional. Initial value for proportional error. Default is 0.1.}

\item{input_add.err}{Optional. Initial value for log-normal distributed additive error. If not provided, it will be set based on observed data.}

\item{input_eta.cl}{Optional. Initial value for ETA on clearance. Default is 0.1.}

\item{input_eta.vc}{Optional. Initial value for ETA on Vc. Default is 0.1.}

\item{input_eta.vp}{Optional. Initial value for ETA on Vp. Default is 0.1.}

\item{input_eta.vp2}{Optional. Initial value for ETA on Vp2. Default is 0.1.}

\item{input_eta.q}{Optional. Initial value for ETA on Q. Default is 0.1.}

\item{input_eta.q2}{Optional. Initial value for ETA on Q2. Default is 0.1.}

\item{input_eta.vmax}{Optional. Initial value for ETA on Vmax. Default is 0.1.}

\item{input_eta.km}{Optional. Initial value for ETA on Km. Default is 0.1.}

\item{input_etarsquare}{Optional. Initial value for the square of the correlation coefficient for ETA. Default is 0.2.}
}
\value{
None. This function updates global variables.
}
\description{
Sets initial values for various model parameters based on provided data and initial estimates.
It updates global variables with the specified or default initial values.
}
\examples{
\dontrun{

library(nlmixr2autoinit)
dat<-pheno_sd
inits.out<-getppkinit(dat = dat,runnpd = 0)
autosetinit(dat = dat,
           inits= inits.out$Recommended_initial_estimates)
}

}
