% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.pop.R
\name{create.pop}
\alias{create.pop}
\title{Generate initial population for genetic algorithm}
\usage{
create.pop(npopsize, nbits)
}
\arguments{
\item{npopsize}{An integer representing the number of individuals (chromosomes) in the population.}

\item{nbits}{An integer representing the number of bits in each chromosome.}
}
\value{
A matrix of size \code{npopsize} by \code{nbits}, where each element is either 0 or 1.
}
\description{
Generates the initial population of chromosomes for a genetic algorithm.
Each chromosome is represented as a binary string with bits that are either 0 or 1.
}
\examples{
# Create an initial population of 10 chromosomes, each with 12 bits
create.pop(npopsize = 10, nbits = 12)

}
