% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.code.R
\name{de.code}
\alias{de.code}
\title{Revise invalid model codes to valid models (GA)}
\usage{
de.code(string, search.space)
}
\arguments{
\item{string}{A binary string representing the model configuration.}

\item{search.space}{An integer indicating the search space.}
}
\value{
A numeric vector representing the revised model string.
}
\description{
Revise invalid model codes to ensure strings generated from GA are valid based on the specified rules.
}
\examples{
\dontrun{
string <- c(0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 1)
de.code(string,1)
}

}
