% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.mod.create.iv.mm.R
\name{ex.mod.create.iv.mm}
\alias{ex.mod.create.iv.mm}
\title{Generate model files for population pharmacokinetic modelling}
\usage{
ex.mod.create.iv.mm(
  modi,
  cmpt.iv,
  eta.km,
  eta.vc,
  eta.vp,
  eta.vp2,
  eta.q,
  eta.q2,
  mm,
  mcorr,
  rv
)
}
\arguments{
\item{modi}{The current model number.}

\item{cmpt.iv}{The number of compartments, represented by a categorical variable:
\itemize{
\item \code{cmpt.iv = 1}: One compartment model
\item \code{cmpt.iv = 2}: Two compartment model
\item \code{cmpt.iv = 3}: Three compartment model
}}

\item{eta.km}{A binary value indicating the presence (1) or absence (0) of random effects for the Michaelis-Menten constant (Km).}

\item{eta.vc}{A binary value indicating the presence (1) or absence (0) of random effects for the central volume of distribution (Vc).}

\item{eta.vp}{A binary value indicating the presence (1) or absence (0) of random effects for the peripheral volume of distribution (Vp).}

\item{eta.vp2}{A binary value indicating the presence (1) or absence (0) of random effects for the second peripheral volume of distribution (Vp2).}

\item{eta.q}{A binary value indicating the presence (1) or absence (0) of random effects for the intercompartmental clearance (Q).}

\item{eta.q2}{A binary value indicating the presence (1) or absence (0) of random effects for the second intercompartmental clearance (Q2).}

\item{mm}{A parameter representing the Michaelis-Menten elimination.}

\item{mcorr}{A parameter representing the existence of correlation between parameters.}

\item{rv}{A categorical variable representing the type of residual error:
\itemize{
\item \code{rv = 1}: Additive error
\item \code{rv = 2}: Proportional error
\item \code{rv = 3}: Combined error
}}
}
\value{
None. The function creates a model file.
}
\description{
Generate a model file based on the specified parameters for use in automated modelling,
and the model's initial estimates, residual errors, and differential equations based on the
number of compartments and the presence of random effects, correlations, and Michaelis-Menten elimination.
}
\examples{
\dontrun{
ex.mod.create.iv.mm(modi = 1, cmpt.iv = 2, eta.km = 1, eta.vc = 1, eta.vp = 1, eta.vp2 = 0,
                    eta.q = 1, eta.q2 = 0, mm = 1, mcorr = 1, rv = 3)
}

}
