% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.mod.create.R
\name{ga.mod.create}
\alias{ga.mod.create}
\title{Create chromosome string of parameters}
\usage{
ga.mod.create(
  modi,
  cmpt.iv1,
  cmpt.iv2,
  eta.km,
  eta.vc,
  eta.vp,
  eta.vp2,
  eta.q,
  eta.q2,
  mm,
  mcorr,
  rv1,
  rv2
)
}
\arguments{
\item{modi}{The current model number.}

\item{cmpt.iv1}{A binary value representing the first part of the number of compartments.}

\item{cmpt.iv2}{A binary value representing the second part of the number of compartments.
\itemize{
\item \code{cmpt.iv1 = 0, cmpt.iv2 = 0}: One compartment
\item \code{cmpt.iv1 = 0, cmpt.iv2 = 1}: One compartment
\item \code{cmpt.iv1 = 1, cmpt.iv2 = 0}: Two compartments
\item \code{cmpt.iv1 = 1, cmpt.iv2 = 1}: Three compartments
}}

\item{eta.km}{A binary value indicating the presence (1) or absence (0) of random effects for the Michaelis-Menten constant (Km).}

\item{eta.vc}{A binary value indicating the presence (1) or absence (0) of random effects for the central volume of distribution (Vc).}

\item{eta.vp}{A binary value indicating the presence (1) or absence (0) of random effects for the peripheral volume of distribution (Vp).}

\item{eta.vp2}{A binary value indicating the presence (1) or absence (0) of random effects for the second peripheral volume of distribution (Vp2).}

\item{eta.q}{A binary value indicating the presence (1) or absence (0) of random effects for the intercompartmental clearance (Q).}

\item{eta.q2}{A binary value indicating the presence (1) or absence (0) of random effects for the second intercompartmental clearance (Q2).}

\item{mm}{A binary value indicating the presence (1) or absence (0) of Michaelis-Menten elimination.}

\item{mcorr}{A binary value indicating the presence (1) or absence (0) of correlation between subjects}

\item{rv1}{A binary value representing the first part of the residual error.}

\item{rv2}{A binary value representing the second part of the residual error.
\itemize{
\item \code{rv1 = 0, rv2 = 0}: Additive residual error
\item \code{rv1 = 0, rv2 = 1}: Additive residual error
\item \code{rv1 = 1, rv2 = 0}: Proportional residual error
\item \code{rv1 = 1, rv2 = 1}: Combined residual error
}}
}
\value{
A vector containing the assembled string of model parameters.
}
\description{
Converts binary codes into categorical codes and assembles them into a string
of model parameters for use in a genetic algorithm. It applies specific rules for the conversion
of input variables and calls the external model creation function \code{ex.mod.create.iv.mm}.
}
\examples{
# Example usage:
ga.mod.create(modi=1001,cmpt.iv1 = 1, cmpt.iv2 = 1, eta.km = 0, eta.vc = 0,
                            eta.vp = 0, eta.vp2 = 1, eta.q = 0, eta.q2 = 0,
                            mm = 1, mcorr = 1, rv1 = 1, rv2 = 0)

}
