% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gacores.R
\name{ga.mutation}
\alias{ga.mutation}
\title{Mutation in Genetic Algorithm}
\usage{
ga.mutation(children.all, prob.mutation, npopsize, nbits)
}
\arguments{
\item{children.all}{A matrix representing the population of children chromosomes.}

\item{prob.mutation}{The probability of mutation for each bit in the chromosome.}

\item{npopsize}{An integer specifying the size of the population.}

\item{nbits}{An integer specifying the number of bits in each chromosome.}
}
\value{
A matrix representing the population of children chromosomes after mutation.
}
\description{
Perform a mutation operation on a population of chromosomes in a genetic algorithm.
Each bit in the chromosome has a chance to mutate based on a predefined mutation probability.
}
\examples{
\dontrun{
children.all <- matrix(sample(0:1, 120, replace = TRUE), nrow = 10)
prob.mutation <- 0.1
npopsize <- 10
nbits <- 12
mutated_population <- ga.mutation(children.all, prob.mutation, npopsize, nbits)
print(mutated_population)
}

}
