% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.operator.R
\name{ga.operator}
\alias{ga.operator}
\title{Run genetic algorithm (GA) for automated model selection}
\usage{
ga.operator(
  dat,
  npopsize,
  max.iter,
  prob.crossover,
  prob.mutation,
  nlocal.search,
  sig.diff,
  search.space,
  no.cores,
  thetalower,
  crse,
  cshrink,
  cadd,
  cprop,
  ccorr,
  penalty.type,
  penalty.value,
  foldername,
  filename,
  ...
)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data to be used for model fitting.}

\item{npopsize}{An integer specifying the size of the population.}

\item{max.iter}{An integer specifying the maximum number of iterations for the genetic algorithm.}

\item{prob.crossover}{The probability of crossover in the genetic algorithm.}

\item{prob.mutation}{The probability of mutation in the genetic algorithm.}

\item{nlocal.search}{The frequency of local exhaustive search iterations.}

\item{sig.diff}{The significance difference threshold for ranking.}

\item{search.space}{An integer specifying the search space type (1 for standard).}

\item{no.cores}{The number of CPU cores to use for parallel processing.}

\item{thetalower}{A named vector specifying lower bounds for model parameters.}

\item{crse}{Constraint value for the relative standard error (RSE).}

\item{cshrink}{Constraint value for the shrinkage.}

\item{cadd}{Constraint value for the additive error.}

\item{cprop}{Constraint value for the proportional error.}

\item{ccorr}{Constraint value for the correlation between parameters.}

\item{penalty.type}{The type of penalty to apply during the fitness evaluation:
\itemize{
\item \code{penalty.type = 1}: Considers constraints on RSE.
\item \code{penalty.type = 2}: Considers constraints on RSE and shrinkage.
\item \code{penalty.type = 3}: Considers constraints on RSE, shrinkage, and sigma values (variance of residual error).
}}

\item{penalty.value}{The value of the penalty to apply during the fitness evaluation.}

\item{foldername}{The name of the folder for output storage. Defaults to "test".}

\item{filename}{The name of the output file. Defaults to "test".}

\item{...}{Additional arguments passed to the model fitting function.}
}
\value{
A matrix representing the best solution found by the genetic algorithm.
}
\description{
Perform a genetic algorithm to select the best model for pharmacokinetic modelling
}
\examples{
\dontrun{
library(nlmixr2autoinit)
library(nlmixr2autoGA)
d1<-pheno_sd
inits.out<-getppkinit(dat = d1,runnpd = 0)
autosetinit(dat = d1,
           inits= inits.out$Recommended_initial_estimates)
best.model<-ga.operator(dat = d1,
                        npopsize = 8,
                        max.iter = 6,
                        prob.crossover = 0.8,
                        prob.mutation = 0.2,
                        nlocal.search = 3,
                        sig.diff = 1,
                        search.space = 1,
                        no.cores = 4,
                        thetalower=c(vp=1,
                                     vp2=1),

                        control = saemControl(
                          seed = 1234,
                          print = 5,
                          nBurn = 20,
                          nEm = 30,
                          logLik = T,
                         rxControl = rxControl(cores = 4,
                                                maxsteps =70000)),
                        table=tableControl(cwres=T),
                        filename = "test",
                        foldername = "test"
)
best.model
}

}
