% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gacores.R
\name{ga.sel.tournament}
\alias{ga.sel.tournament}
\title{Tournament selection for genetic algorithm}
\usage{
ga.sel.tournament(data.pop, population, npopsize, nbits)
}
\arguments{
\item{data.pop}{A data frame containing the current population with their fitness values and rankings.}

\item{population}{A matrix representing the current population of chromosomes.}

\item{npopsize}{An integer specifying the size of the population.}

\item{nbits}{An integer specifying the number of bits in each chromosome.}
}
\value{
A matrix representing the selected population for the next generation.
}
\description{
Perform tournament selection to choose individuals for the next generation
in a genetic algorithm.
}
\examples{
data.pop <- data.frame(fitness = runif(10), rank = rank(runif(10)))
population <- matrix(sample(0:1, 100, replace = TRUE), nrow = 10)
npopsize <- 10
nbits <- 10
ga.sel.tournament(data.pop, population, npopsize, nbits)

}
