% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabu.operator.R
\name{print.tabuOperatorResult}
\alias{print.tabuOperatorResult}
\title{Print method for tabu operator results}
\usage{
\method{print}{tabuOperatorResult}(x, ...)
}
\arguments{
\item{x}{A "tabuOperatorResult" object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print tabu operator results.
}
\seealso{
\code{\link{tabu.operator}}
}
