% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.code.R
\name{read.code2}
\alias{read.code2}
\title{Generate pharmacokinetic model name based on selected codes \link{2}}
\usage{
read.code2(search.space, sel.best.code)
}
\arguments{
\item{search.space}{An integer indicating the search space.}

\item{sel.best.code}{A numeric vector containing the selected best codes. If \code{search.space} is 1, it should have 12 elements representing different components and parameters.}
}
\value{
A character string representing the model name.
}
\description{
Generate a model name based on the selected best code and specified search space for categorical coding algorithm
}
\details{
The function uses the values in \code{sel.best.code} to determine the components and parameters of the model name. The model name is constructed based on the following:
\itemize{
\item \code{cmpt}: Determine the number of compartments (1, 2, or 3).
\item \code{eta.km}, \code{eta.vc}, \code{eta.vp}, \code{eta.vp2}, \code{eta.q}, and \code{eta.q2}: Determine the presence of various random effects.
\item \code{mm}: Determine the elimination method (first-order or Michaelis-Menten).
\item \code{mcorr}: Determine whether correlation is present.
\item \code{rv}: Determine the residual variance type (additive, proportional, or combined).
}
}
\examples{
sel.best.code <- c(1, 1, 0, 0, 0, 0, 0, 0, 0, 1)
model_name <- read.code(1, sel.best.code)
print(model_name)

}
