% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.mod.run.R
\name{sf.mod.run}
\alias{sf.mod.run}
\title{Run and evaluate a model (stepwise)}
\usage{
sf.mod.run(modi, string, ...)
}
\arguments{
\item{modi}{The current model number.}

\item{string}{A binary string representing the genetic algorithm's chromosome.}

\item{...}{Additional arguments to be passed to the \code{nlmixr2} function.}
}
\value{
A numeric value representing the fitness of the model.
}
\description{
Conduct a model run and evaluate its fitness within the context of a genetic algorithm.
Converts binary strings to model parameters, fits the model using \code{nlmixr2},
and evaluates the model's performance based on \code{fitness}.
}
\examples{
\dontrun{
dat <- pheno_sd
inits.out<-getppkinit(dat = dat)
autosetinit(dat = dat,
           inits= inits.out$Recommended_initial_estimates)
string <- c(1, 0, 0, 0, 0, 0, 0, 0, 0, 1)
fitness <- sf.mod.run(modi=1, r = 1, dat = dat, search.space = 1, string = string,
                      crse = 20, cshrink = 30, lbcl = 1, lbvc = 1, lbvp = 1,
                      lbq = 0.01, lbvp2 = 1, lbq2 = 0.01, cadd = 0.1, cprop = 0.01,
                      ccorr = 0.2, penalty.type = 3, penalty.value = 10000,control = saemControl(logLik = T))
}

}
