% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gacores.R
\name{ga.crossover}
\alias{ga.crossover}
\title{Crossover in genetic algorithm}
\usage{
ga.crossover(sel.population, prob.crossover, npopsize, nbits, bit.names)
}
\arguments{
\item{sel.population}{A matrix representing the selected population of chromosomes.}

\item{prob.crossover}{The probability of crossover in the genetic algorithm.}

\item{npopsize}{An integer specifying the size of the population.}

\item{nbits}{An integer specifying the number of bits in each chromosome.}

\item{bit.names}{A vector of character strings representing the names of the bits.}
}
\value{
A matrix representing the population of children chromosomes after crossover.
}
\description{
Perform a crossover operation on a selected population in a genetic algorithm.
Use a predefined probability to decide whether to perform crossover on pairs of parent chromosomes
and generates children chromosomes by combining segments of the parent chromosomes.
}
\examples{
sel.population <- matrix(sample(0:1, 100, replace = TRUE), nrow = 10)
prob.crossover <- 0.8
npopsize <- 10
nbits <- 12
ga.crossover(sel.population, prob.crossover, npopsize, nbits, bit.names)

}
