% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabu.neighbor.search.R
\name{generate_neighbors_df}
\alias{generate_neighbors_df}
\title{Generate neighbors data frame}
\usage{
generate_neighbors_df(current_string)
}
\arguments{
\item{current_string}{A named vector or list representing the current state,
where names correspond to variable names and values correspond to their current values.}
}
\value{
A DataFrame containing all possible neighboring states.
}
\description{
Generate a data frame containing all possible neighbor models of a given current model string.
}
\examples{
current_string <- list(cmpt.iv = 2, eta.km = 0, eta.vc = 1, eta.vp = 0, 
                      eta.vp2 = 1, eta.q = 0, eta.q2 = 1, mm = 0, 
                      mcorr = 1, rv = 2)
neighbors_df <- generate_neighbors_df(current_string)
print(neighbors_df)

}
