% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mod.lst.R
\name{get.mod.lst}
\alias{get.mod.lst}
\title{Extract model parameters and statistics}
\usage{
get.mod.lst(fit.s, modi)
}
\arguments{
\item{fit.s}{A fitted \code{nlmixr2} model object.}

\item{modi}{The current model number.}
}
\value{
A data frame containing the extracted model parameters and statistics.
}
\description{
Extract model parameters and statistics from a fitted \code{nlmixr2} model object.
}
\examples{
\dontrun{

pheno <- function() {
  ini({
    tcl <- log(0.008) # typical value of clearance
    tv <-  log(0.6)   # typical value of volume
    eta.cl + eta.v ~ c(1,
                       0.01, 1) ## cov(eta.cl, eta.v), var(eta.v)
    add.err <- 0.1    # residual variability
  })
  model({
    cl <- exp(tcl + eta.cl) # individual value of clearance
    v <- exp(tv + eta.v)    # individual value of volume
    ke <- cl / v            # elimination rate constant
    d/dt(A1) = - ke * A1    # model differential equation
    cp = A1 / v             # concentration in plasma
    cp ~ add(add.err)       # define error model
  })
}

# Fit the model using nlmixr2
fit <- nlmixr2(one.compartment, pheno_sd, est="saem", saemControl(print=0))

# Extract model results
model_results <- get.mod.lst(fit,1)
print(model_results)
}

}
