% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aco.cores.R
\name{p.calculation}
\alias{p.calculation}
\title{Calculate selection probabilities for nodes}
\usage{
p.calculation(search.space, node.list.cal)
}
\arguments{
\item{search.space}{An integer specifying the search space type (1 for standard).}

\item{node.list.cal}{A data frame containing the list of nodes with their pheromone levels and other attributes.}
}
\value{
A data frame representing the updated node list with calculated selection probabilities.
}
\description{
Calculate the selection probabilities for nodes in the ant colony optimization
}
\examples{
\dontrun{
# Example usage:
search.space <- 1
no.nodes <- 22
initial.phi <- 1
node.list.cal <- data.frame(
  travel = 1,
  node.no = seq(1, no.nodes, 1),
  local.node.no = c(seq(1, 3, 1), seq(0, 1, 1), seq(0, 1, 1), seq(0, 1, 1), seq(0, 1, 1),
                    seq(0, 1, 1), seq(0, 1, 1), seq(0, 1, 1), seq(0, 1, 1), seq(1, 3, 1)),
  node.names = c("1Cmpt", "2Cmpt", "3Cmpt", "eta.vp2.no", "eta.vp2.yes", "eta.q2.no", "eta.q2.yes",
                 "eta.vp.no", "eta.vp.yes", "eta.q.no", "eta.q.yes", "eta.vc.no", "eta.vc.yes",
                 "mm.no", "mm.yes", "eta.km.no", "eta.km.yes", "mcorr.no", "mcorr.yes",
                 "add", "prop", "comb"),
  node.group = c(rep(1, 3), sort(rep(seq(2, 9, 1), 2)), rep(10, 3)),
  phi = seq(1,22),
  delta_phi = rep(0, no.nodes),
  p = c(rep(round(1 / 3, 3), 3), rep(0.5, 16), rep(round(1 / 3, 3), 3))
)
updated.node.list.cal <- p.calculation(search.space, node.list.cal)
}

}
