% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{param.bounds}
\alias{param.bounds}
\title{Construct default parameter bounds for model fitness evaluation}
\usage{
param.bounds(
  theta = list(lower = NULL, upper = NULL),
  omega = list(lower = NULL, upper = NULL),
  sigma = list(add = list(lower = NULL, upper = Inf), prop = list(lower = 0.05, upper =
    Inf)),
  correlation = list(lower = 0.1, upper = Inf)
)
}
\arguments{
\item{theta}{List of lower and upper bounds for fixed-effect parameters (theta)}

\item{omega}{List of lower bounds for BSV parameters (omega)}

\item{sigma}{List with named lower bounds for residual error (additive and proportional)}

\item{correlation}{List of lower and upper bounds for interindividual correlation}
}
\value{
A structured list containing parameter bounds
}
\description{
Construct default parameter bounds for model fitness evaluation
}
