% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aco.operator.R
\name{print.acoOperatorResult}
\alias{print.acoOperatorResult}
\title{Print method for ACO operator results}
\usage{
\method{print}{acoOperatorResult}(x, ...)
}
\arguments{
\item{x}{An "acoOperatorResult" object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print ACO operator results.
}
\seealso{
\code{\link{aco.operator}}
}
\author{
Zhonghui Huang
}
