% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.operator.R
\name{print.sfOperatorResult}
\alias{print.sfOperatorResult}
\title{Print method for sfOperatorResult objects}
\usage{
\method{print}{sfOperatorResult}(x, ...)
}
\arguments{
\item{x}{An object of class 'sfOperatorResult'.}

\item{...}{Further arguments passed to or from other methods
(currently unused).}
}
\value{
Invisibly returns x.
}
\description{
Defines a custom print method for objects of class
'sfOperatorResult'.
}
