% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.method.R
\name{rank_new}
\alias{rank_new}
\title{Ranking with significance difference threshold}
\usage{
rank_new(x1, sig.diff)
}
\arguments{
\item{x1}{A numeric vector to be ranked.}

\item{sig.diff}{A numeric value specifying the significance difference threshold. Values within this
threshold are considered equal and receive the same rank.}
}
\value{
A numeric vector representing the adjusted ranks of the input values.
}
\description{
Performs a custom ranking of a numeric vector,and ajusts the ranks of values that
differ by less than a specified threshold, ensuring they receive the same rank.
}
\examples{
\dontrun{
x1 <- c(10, 20, 20.5, 30)
sig.diff <- 1
ranked_list <- rank_new(x1, sig.diff)
print(ranked_list)
}

}
