% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabu.operator.R
\name{tabu.operator}
\alias{tabu.operator}
\title{Run Tabu search for automated pharmacokinetic model selection}
\usage{
tabu.operator(
  dat,
  tabu.duration,
  max.round,
  start.point,
  search.space,
  no.cores,
  thetalower,
  crse,
  cshrink,
  cadd,
  cprop,
  ccorr,
  penalty.type,
  penalty.value,
  foldername,
  filename,
  ...
)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data to be used for model fitting.}

\item{tabu.duration}{Integer specifying the number of iterations a target remains tabu list.}

\item{max.round}{Integer specifying the maximum number of rounds for the search.}

\item{start.point}{Named vector specifying a model as starting points for the search.}

\item{search.space}{An integer specifying the search space type (1 for standard).}

\item{no.cores}{The number of CPU cores to use for parallel processing.}

\item{thetalower}{A named vector specifying lower bounds for model parameters.}

\item{crse}{Constraint value for the relative standard error (RSE).}

\item{cshrink}{Constraint value for the shrinkage.}

\item{cadd}{Constraint value for the additive error.}

\item{cprop}{Constraint value for the proportional error.}

\item{ccorr}{Constraint value for the correlation between parameters.}

\item{penalty.type}{The type of penalty to apply during the fitness evaluation:
\itemize{
\item \code{penalty.type = 1}: Considers constraints on RSE.
\item \code{penalty.type = 2}: Considers constraints on RSE and shrinkage.
\item \code{penalty.type = 3}: Considers constraints on RSE, shrinkage, and sigma values (variance of residual error).
}}

\item{penalty.value}{The value of the penalty to apply during the fitness evaluation.}

\item{foldername}{The name of the folder for output storage. Defaults to "test".}

\item{filename}{The name of the output file. Defaults to "test".}

\item{...}{Additional arguments passed to the model fitting function.}
}
\value{
A list include the best solution and Tabu running history
}
\description{
Perform Tabu search to select the best model for pharmacokinetic modelling
}
\examples{
\dontrun{

d1<-pheno_sd
inits.out<-getppkinit(dat = d1,runnpd = 0)
autosetinit(dat = d1,
            inits= inits.out$Recommended_initial_estimates)
tabu.operator(dat = d1,
              tabu.duration=2,
              max.round=4,
              search.space = 1,
              no.cores = 4,
              thetalower=c(vp=1,
                           vp2=1),
              control = saemControl(
                seed = 1234,
                print = 5,
                nBurn = 20,
                nEm = 30,
                logLik = T,
                rxControl = rxControl(cores = 4,
                                      maxsteps =70000)),
               table=tableControl(cwres=T),
               filename =  "pheno_sd",
               foldername =   "pheno_sd" )


}
}
